+++ cnccode/business_classes/BUActivity.inc.php
@@ -806,7 +806,9 @@ class BUActivity extends Business
             );
             $dbeProblem->updateRow();
             $message = "{$this->dbeUser->getValue(DBEUser::name)} Escalated from {$this->workQueueDescriptionArray[$oldQueueNo]} to {$this->workQueueDescriptionArray[$newQueueNo]}";
-            if ($dbeProblem->getValue(DBEProblem::status) == 'P') {
+            //if ($dbeProblem->getValue(DBEProblem::status) == 'P') 
+            if($reason!='')
+            {
                 $message .= " because of {$reason}";
             }
             $message .= ".";
@@ -10746,10 +10748,10 @@ FROM
      * @param $customerID
      * @return DBEJProblem
      */
-    function getCustomerOpenSR($customerID)
+    function getCustomerOpenSR($customerID,$srNumber=null)
     {
         $dbeJProblem = new DBEJProblem($this);
-        $dbeJProblem->getCustomerOpenRows($customerID);
+        $dbeJProblem->getCustomerOpenRows($customerID,$srNumber);
         return $dbeJProblem;
 
     }